Rename utils/test.py, move utils tests into utils/tests module
